pyproject.toml
[project]
name = "zooplotly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["kaleido==0.2.1", "plotly>=5.24.1", "requests>=2.32.3"]

[dependency-groups]
dev = ["numpy>=2.2.2", "pillow>=11.1.0", "pytest>=8.3.4", "ruff>=0.9.3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/zooplotly"]

[tool.hatch.build.targets.wheel.package-data]
zooplotlib = ["templates/*.json"]
